RAG/项目五_多模态RAG/requirements.txt
# 基础依赖
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pymupdf==1.23.8
markdown==3.5.1
python-dotenv==1.0.0
openai==1.3.0
tiktoken==0.5.1
numpy==1.24.3
pandas==2.0.3

# 多模态特定依赖
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
opencv-python==4.8.1.78
clip-by-openai==1.1.0
sentencepiece==0.1.99
